requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.29
pymysql>=1.1.0
cryptography>=42.0.5
pydantic>=2.7.0
pydantic[email]>=2.7.0
pydantic-settings>=2.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.9
pdfplumber>=0.11.0
httpx>=0.27.0
openai>=1.30.1
google-generativeai>=0.5.2
python-dotenv>=1.0.1
alembic>=1.13.1
